fix: 댓글 삭제 시 대댓글을 포함한 갯수를 투표에서 감소하게 변경

+++ src/main/java/com/salmalteam/salmal/comment/application/CommentService.java
@@ -52,13 +52,14 @@ public class CommentService {
 	public void deleteComment(final Long memberId, final Long commentId) {
 		final Comment comment = getCommentById(commentId);
 		validateDeleteAuthority(comment.getCommenter().getId(), memberId);
-
-		switch (comment.getCommentType()) {
-			case COMMENT:
-				commentRepository.deleteAllRepliesByParentCommentId(commentId);
-			case REPLY:
-				commentRepository.decreaseReplyCount(commentId);
+		if (comment.isComment()) {
+			Long deleteComment = commentRepository.countByParentComment(comment) + 1;
+			commentRepository.deleteAllRepliesByParentCommentId(commentId);
+			voteRepository.decreaseCommentCount(comment.getVote().getId(), deleteComment);
+			commentRepository.delete(comment);
+			return;
 		}
+		commentRepository.decreaseReplyCount(commentId);
 		voteRepository.decreaseCommentCount(comment.getVote().getId());
 		commentRepository.delete(comment);
 	}
